refactor: Consolidate slice tests
